Restructured example applications.
