+++ examples/CMakeLists.txt
@@ -13,6 +13,15 @@ target_link_libraries( cli_hello
   smack_cpp
 )
 
+add_executable( cli_hello_new_type
+    cli_hello_new_type.cpp
+    ${headers}
+)
+
+target_link_libraries( cli_hello_new_type
+  smack_cpp
+)
+
 add_executable( cli_class
     cli_class.cpp
     ${headers}
@@ -22,29 +31,11 @@ target_link_libraries( cli_class
   smack_cpp
 )
 
-add_executable( cli_class_argcv
-  cli_class_argcv.cpp
-  ${headers}
-)
-
-target_link_libraries( cli_class_argcv
-  smack_cpp
-)
-
 add_executable( cli_functional
     cli_functional.cpp
     ${headers}
 )
 
 target_link_libraries( cli_functional
-  smack_cpp
-)
-
-add_executable( cli_functional_argcv
-    cli_functional_argcv.cpp
-    ${headers}
-)
-
-target_link_libraries( cli_functional_argcv
   smack_cpp
 )
